Cleanup integration tests (#1066)
