+++ test/integration/nginxless/nginx_less_mpi_connection_test.go
@@ -3,19 +3,22 @@
 // This source code is licensed under the Apache License, Version 2.0 license found in the
 // LICENSE file in the root directory of this source tree.
 
-package integration
+package nginxless
 
 import (
 	"testing"
+
+	"github.com/nginx/agent/v3/test/integration/utils"
 
 	"github.com/stretchr/testify/assert"
 )
 
 // Verify that the agent sends a connection request to Management Plane even when Nginx is not present
 func TestNginxLessGrpc_Connection(t *testing.T) {
-	teardownTest := setupConnectionTest(t, true, true, "../config/agent/nginx-config-with-max-file-size.conf")
+	teardownTest := utils.SetupConnectionTest(t, true, true,
+		"../../config/agent/nginx-config-with-grpc-client.conf")
 	defer teardownTest(t)
 
-	verifyConnection(t, 1)
+	utils.VerifyConnection(t, 1)
 	assert.False(t, t.Failed())
 }
